Transfer some of the fragment functionality to the activity
