+++ src/main/res/layout/fragment_behavior_progress.xml
@@ -5,33 +5,10 @@
     android:layout_height="match_parent" >
 
     <TextView
-        android:id="@+id/behavior_progress_title"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="0.50"
-        android:layout_gravity="center_vertical"
-        android:gravity="center"
-        android:padding="25dp"
-        android:textSize="25sp"
-        android:textColor="#FFF"
-        android:background="@color/grow_primary_dark"
-        android:text="@string/behavior_progress_title"/>
-
-    <ProgressBar
-        style="?android:attr/progressBarStyle"
-        android:id="@+id/behavior_progress_progress_bar"
-        android:layout_width="wrap_content"
-        android:layout_height="0dp"
-        android:layout_weight="0.1"
-        android:layout_gravity="center"
-        android:layout_marginTop="10dp"
-        android:visibility="invisible" />
-
-    <TextView
         android:id="@+id/behavior_progress_behavior_label"
         android:layout_width="match_parent"
         android:layout_height="0dp"
-        android:layout_weight="0.2"
+        android:layout_weight="0.5"
         android:layout_marginLeft="20dp"
         android:gravity="center_vertical"
         android:textColor="@color/dark_text_color"
@@ -42,47 +19,10 @@
         android:id="@+id/behavior_progress_behavior_title"
         android:layout_width="match_parent"
         android:layout_height="0dp"
-        android:layout_weight="0.2"
+        android:layout_weight="0.5"
         android:layout_marginLeft="20dp"
         android:textColor="@color/grow_primary"
         android:textAppearance="?android:attr/textAppearanceMedium"
         android:padding="10dp" />
 
-    <LinearLayout
-        android:orientation="horizontal"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="right" >
-
-        <ImageButton
-            android:id="@+id/behavior_progress_on_track"
-            android:layout_width="@dimen/fab_size_normal"
-            android:layout_height="@dimen/fab_size_normal"
-            android:src="@drawable/compass_1_n"
-            android:scaleType="fitCenter"
-            android:background="@drawable/circle_primary"
-            android:contentDescription="@string/behavior_progress_on_track"
-            android:layout_margin="10dp" />
-
-        <ImageButton
-            android:id="@+id/behavior_progress_seeking"
-            android:layout_width="@dimen/fab_size_normal"
-            android:layout_height="@dimen/fab_size_normal"
-            android:src="@drawable/compass_5_e"
-            android:scaleType="fitCenter"
-            android:background="@drawable/circle_primary"
-            android:contentDescription="@string/behavior_progress_seeking"
-            android:layout_margin="10dp" />
-
-        <ImageButton
-            android:id="@+id/behavior_progress_off_course"
-            android:layout_width="@dimen/fab_size_normal"
-            android:layout_height="@dimen/fab_size_normal"
-            android:src="@drawable/compass_9_s"
-            android:scaleType="fitCenter"
-            android:background="@drawable/circle_primary"
-            android:contentDescription="@string/behavior_progress_off_course"
-            android:layout_margin="10dp" />
-
-    </LinearLayout>
 </LinearLayout>
